books_rest_api/requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
